+++ app/controllers/ahoy_captain/cities_controller.rb
@@ -10,11 +10,7 @@ module AhoyCaptain
 
     def index
       results = cached(:cities) do
-        visit_query
-          .select("city, country, count(concat(city, region, country)) as count, sum(count(concat(city, region, country))) over() as total_count")
-          .where.not(city: nil)
-          .group("city, region, country")
-          .order(Arel.sql "count(concat(city, region, country)) desc")
+        CityQuery.call(params)
           .limit(limit)
       end