+++ CHANGELOG.md
@@ -10,6 +10,12 @@
   * this was a lazy man's RequestStore for one attribute and instead we'll just send it with the decorator
 * Default period is now 30d, was MTD
 * Simplified the filters
+* All front-facing queries now live in a query object
+  * makes exporting easier 
+* Add CSV export 
+
+### Fixed 
+* Views/visit returning NaN 
 
 ## Version 0.8